+++ lib/multi_selection.dart
@@ -20,6 +20,18 @@ class CupertinoMultiSelection<T> extends StatelessWidget {
   /// Displayed above the options. Usually a [Text] widget.
   final Widget? header;
 
+  /// Displayed below the options. Usually a [Text] widget.
+  final Widget? footer;
+
+  /// The type of the header. If "insetGrouped", the header will be displayed
+  /// large and bold like the iOS notes app. If "base", the header will be displayed
+  /// smaller, like the iOS settings app.
+  final CupertinoListSectionType? headerType;
+
+  /// The type of the footer. If "insetGrouped", no styling will be applied.
+  /// If "base", the footer will be displayed in the style of the iOS settings app.
+  final CupertinoListSectionType? footerType;
+
   /// Determines whether the check mark will be located on the leading (left) or trailing (right) side of the tile.
   final WidgetSelectionLocation checkMarkLocation;
 
@@ -28,6 +40,9 @@ class CupertinoMultiSelection<T> extends StatelessWidget {
     required this.selected,
     this.onChanged,
     this.header,
+    this.footer,
+    this.headerType,
+    this.footerType,
     this.checkMarkLocation = WidgetSelectionLocation.trailing,
     super.key});
 
@@ -35,6 +50,9 @@ class CupertinoMultiSelection<T> extends StatelessWidget {
   Widget build(BuildContext context) {
     return CupertinoSelection(
       header: header,
+      footer: footer,
+      headerType: headerType,
+      footerType: footerType,
       children: children.map((item) => CupertinoSelectionTile<T>(
         title: item.title,
         subtitle: item.subtitle,