+++ CHANGELOG.md
@@ -5,3 +5,9 @@
 ## 0.0.2
 
 * Added showCupertinoSingleSelectionPage and SingleSelectionPage, which provide a simple full screen dialog to select an option, similar to the iOS settings app.
+
+## 0.0.3
+
+* Added footer. 
+* Added EnhancedCupertinoListSection which enables mixing of the "base" and "insetGrouped" styles, enabling the new-style curved lists with the original-style headers and footers, creating an appearance like the iOS settings app.
+* Added headerType and footerType arguments to enable custom combinations of original and new style lists.